fix: deduplicate Backblaze MCP registry tests
